# Allocate the appropriate memory for retrieveIndicesFromBitfield