--- a/beacon-chain/p2p/peers/status.go
+++ b/beacon-chain/p2p/peers/status.go
@@ -493,7 +493,7 @@ func (p *Status) HighestEpoch() uint64 {
 }
 
 func retrieveIndicesFromBitfield(bitV bitfield.Bitvector64) []uint64 {
-	committeeIdxs := []uint64{}
+	committeeIdxs := make([]uint64, 0, bitV.Count())
 	for i := uint64(0); i < 64; i++ {
 		if bitV.BitAt(i) {
 			committeeIdxs = append(committeeIdxs, i)