**What type of PR is this?**

Other

**What does this PR do? Why is it needed?**

Reduces memory allocations by 50% in bytes and 6 allocs per operation.

Before

```
goos: linux
goarch: amd64
pkg: github.com/prysmaticlabs/prysm/beacon-chain/p2p/peers
Benchmark_retrieveIndicesFromBitfield-8          2296381               499 ns/op            1016 B/op          7 allocs/op
PASS
ok      github.com/prysmaticlabs/prysm/beacon-chain/p2p/peers   2.706s
```


After

```
goos: linux
goarch: amd64
pkg: github.com/prysmaticlabs/prysm/beacon-chain/p2p/peers
Benchmark_retrieveIndicesFromBitfield-8          4872231               247 ns/op             512 B/op          1 allocs/op
PASS
ok      github.com/prysmaticlabs/prysm/beacon-chain/p2p/peers   2.917s
```

**Which issues(s) does this PR fix?**

N/A - general memory allocation problem

**Other notes for review**

This is pretty much a 1 line change in production code. Added a benchmark to sanity check and quantify the improvement.